NBCR-2023-001: Remove URI definition from crypto-coin-identity UR type
Defining both the UR type and the URI in the same document leads to
confusion since both formats are used for different purposes.
While the URI is an easily readable coin identifier, specifically fitted
to be defined in an internal file stored in different wallets (both
offline signers and watch-only wallets), the UR type provides
similar information in a compact format to be encoded inside a QR code.